optimize API according to uvwasi

+++ 02assemblyscript/01wasi/assembly/wasi_snapshot_preview1.ts
@@ -3,18 +3,18 @@ export declare function fd_write(
     iov_array_ptr: usize,
     iov_array_length: u32,
     nwritten: usize,
-): i32;
+): u16;
 
 export declare function path_open(
     dirfd: u32,
     dirflags: u32,
     path_ptr: usize,
     path_len: u32,
-    o_flags: u32,
+    o_flags: u16,
     fs_rights_base: u64,
     fs_rights_inheriting: u64,
-    fs_flags: u32,
+    fs_flags: u16,
     fdPtr: usize
-): i32;
+): u16;
 
-export declare function fd_close(fd: u32): i32;
+export declare function fd_close(fd: u32): u16;
